+++ popolo/__init__.py
@@ -1,3 +1,3 @@
 "A Django-based implementation of the Popolo data specifications."
 
-__version__ = '1.0.0'
+__version__ = '1.0.1'