+++ popolo/admin.py
@@ -1,4 +1,5 @@
 from django.contrib import admin
+from django.utils.translation import ugettext_lazy as _
 
 try:
     from django.contrib.contenttypes.admin import GenericTabularInline
@@ -20,24 +21,24 @@ class PersonAdmin(admin.ModelAdmin):
         (None, {
             'fields': ('name', 'gender', 'birth_date', 'death_date')
         }),
-        ('Biography', {
+        (_('Biography'), {
             'classes': ('collapse',),
             'fields': ('summary', 'image', 'biography')
         }),
-        ('Honorifics', {
+        (_('Honorifics'), {
             'classes': ('collapse',),
             'fields': ('honorific_prefix', 'honorific_suffix')
         }),
-        ('Demography', {
+        (_('Demography'), {
             'classes': ('collapse',),
             'fields': ('national_identity',)
         }),
-        ('Special Names', {
+        (_('Special Names'), {
             'classes': ('collapse',),
             'fields': ('family_name', 'given_name',
                        'additional_name', 'patronymic_name', 'sort_name')
         }),
-        ('Advanced options', {
+        (_('Advanced options'), {
             'classes': ('collapse',),
             'fields': ('start_date', 'end_date')
         }),
@@ -52,8 +53,8 @@ class OrganizationMembersInline(MembershipInline):
 
 
 class OrganizationOnBehalfInline(MembershipInline):
-    verbose_name = "Proxy member"
-    verbose_name_plural = "Members acting on behalf of this organization"
+    verbose_name = _("Proxy member")
+    verbose_name_plural = _("Members acting on behalf of this organization")
     fk_name = 'on_behalf_of'
 
 
@@ -63,7 +64,7 @@ class PostAdmin(admin.ModelAdmin):
         (None, {
             'fields': ('label', 'role', 'start_date', 'end_date')
         }),
-        ('Details', {
+        (_('Details'), {
             'classes': ('collapse',),
             'fields': ('other_label', 'area', 'organization')
         }),
@@ -80,13 +81,17 @@ class OrganizationAdmin(admin.ModelAdmin):
         (None, {
             'fields': ('name', 'founding_date', 'dissolution_date')
         }),
-        ('Details', {
+        (_('Details'), {
             'classes': ('collapse',),
             'fields': ('summary', 'image', 'description')
         }),
-        ('Advanced options', {
+        (_('Advanced options'), {
             'classes': ('collapse',),
-            'fields': ('classification', 'start_date', 'end_date')
+            'fields': (
+                'classification',
+                'parent', 'area',
+                'start_date', 'end_date'
+            )
         }),
     )
     inlines = generics.BASE_INLINES + [
@@ -94,7 +99,34 @@ class OrganizationAdmin(admin.ModelAdmin):
         OrganizationOnBehalfInline
     ]
 
+class AreaI18NameInlineAdmin(admin.StackedInline):
+    extra = 0
+    model = models.AreaI18Name
+
+class AreaAdmin(admin.ModelAdmin):
+    model = models.Area
+    fieldsets = (
+        (None, {
+            'fields': (
+                'name', 'identifier', 'classification', 'parent'
+            )
+        }),
+        (_('Details'), {
+            'classes': ('collapse',),
+            'fields': (
+                'classification',
+                'inhabitants',
+            )
+        }),
+    )
+    inlines = [
+        AreaI18NameInlineAdmin,
+        generics.SourceAdmin
+
+    ]
 
 admin.site.register(models.Post, PostAdmin)
 admin.site.register(models.Person, PersonAdmin)
 admin.site.register(models.Organization, OrganizationAdmin)
+admin.site.register(models.Area, AreaAdmin)
+admin.site.register(models.Language)