+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.0.1] - 2017-09-14
+
+### Added
+- italian translation in models and admin
+
+### Changed
+- Area and inline AreaI18Names admin classes
+- models code readability increased
+
 ## [1.0.0] - 2017-09-13
 
 ### Added